orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aspire-orchestrator"
version = "0.1.0"
description = "Aspire LangGraph Orchestrator — the Single Brain (Law #1)"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.3.0,<0.4",
    "langchain-core>=0.3.0,<0.4",
    "langchain-openai>=0.3.0,<0.4",
    "openai>=1.60.0",
    "nemoguardrails>=0.11.0,<0.12",
    "presidio-analyzer>=2.2.0,<3.0",
    "presidio-anonymizer>=2.2.0,<3.0",
    "httpx>=0.28.0,<0.29",
    "pydantic>=2.10.0,<3.0",
    "fastapi>=0.115.0,<0.116",
    "uvicorn[standard]>=0.34.0,<0.35",
    "pyyaml>=6.0,<7.0",
    "prometheus-client>=0.21.0,<1.0",
    "supabase>=2.0.0,<3.0",
    "jsonschema>=4.23.0,<5.0",
    "PyJWT>=2.9.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0",
    "pytest-asyncio>=0.25.0,<0.26",
    "pytest-cov>=6.0,<7.0",
    "httpx",  # for TestClient
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src/aspire_orchestrator"]

[tool.hatch.build.targets.wheel.force-include]
"src/aspire_orchestrator/config/policy_matrix.yaml" = "aspire_orchestrator/config/policy_matrix.yaml"
"src/aspire_orchestrator/config/skill_pack_manifests.yaml" = "aspire_orchestrator/config/skill_pack_manifests.yaml"
"src/aspire_orchestrator/config/tax_rules" = "aspire_orchestrator/config/tax_rules"
"src/aspire_orchestrator/schemas" = "aspire_orchestrator/schemas"
